+++ omniisaacgymenvs/utils/config_utils/sim_config.py
@@ -49,6 +49,8 @@ class SimConfig():
         if self._config["headless"] == True and not self._sim_params["enable_cameras"] and not self._config["enable_livestream"]:
             self._sim_params["use_flatcache"] = False
             self._sim_params["enable_viewport"] = False
+        else:
+            self._sim_params["enable_viewport"] = True
 
         if self._sim_params["disable_contact_processing"]:
             carb.settings.get_settings().set_bool("/physics/disableContactProcessing", True)