+++ omniisaacgymenvs/utils/config_utils/sim_config.py
@@ -46,11 +46,14 @@ class SimConfig():
         if self._config["test"] == True:
             self._sim_params["enable_scene_query_support"] = True
 
+        from omni.isaac.core.utils.extensions import enable_extension
         if self._config["headless"] == True and not self._sim_params["enable_cameras"] and not self._config["enable_livestream"]:
             self._sim_params["use_flatcache"] = False
             self._sim_params["enable_viewport"] = False
         else:
             self._sim_params["enable_viewport"] = True
+            enable_extension("omni.kit.viewport.bundle")
+        enable_extension("omni.replicator.isaac")
 
         if self._sim_params["disable_contact_processing"]:
             carb.settings.get_settings().set_bool("/physics/disableContactProcessing", True)